requirements.txt
streamlit
openai-whisper
torch
torchvision
setuptools